implement deleting images feature

+++ app/Http/Controllers/AdminUsersController.php
@@ -138,8 +138,9 @@ class AdminUsersController extends Controller
 
         $user = User::findOrFail($id);
 
-        unlink(public_path() . "/images/" .  $user->photo->file);
-
+        if($user->photo) {
+            unlink(public_path() . "/images/" . $user->photo->file);
+        }
         $user->delete();
 
         //Session::flash('deleted_user', 'The user has been deleted');
